requirements.txt
mysql-connector-python==8.0.30
numpy==1.23.3
scikit-learn==1.1.2
